+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.2.4 - Unreleased
 
+### Fixed
+* Fixed `cats_per_img` in kwcoco stats.
+
 
 ## Version 0.2.3 - Released 2021-05-13